extract signing and publication routes
